adding TLS flag and edit logic
