+++ pyrit/prompt_target/http_target/http_target.py
@@ -1,15 +1,15 @@
 # Copyright (c) Microsoft Corporation.
 # Licensed under the MIT license.
 
-import logging
 import json
-from typing import Callable, Union
+import logging
+import re
 import requests
-from pyrit.prompt_target import PromptTarget
+from typing import Callable, Union
+
 from pyrit.memory import MemoryInterface
 from pyrit.models import construct_response_from_request, PromptRequestPiece, PromptRequestResponse
-import re
-
+from pyrit.prompt_target import PromptTarget
 
 logger = logging.getLogger(__name__)
 
@@ -22,6 +22,7 @@ class HTTPTarget(PromptTarget):
         prompt_regex_string (str): the placeholder for the prompt
             (default is {PROMPT}) which will be replaced by the actual prompt.
             make sure to modify the http request to have this included, otherwise it will not be properly replaced!
+        use_tls: (bool): whether to use TLS or not. Default is True
         callback_function (function): function to parse HTTP response.
             These are the customizable functions which determine how to parse the output
         memory : memory interface
@@ -31,6 +32,7 @@ class HTTPTarget(PromptTarget):
         self,
         http_request: str = None,
         prompt_regex_string: str = "{PROMPT}",
+        use_tls: bool = True,
         callback_function: Callable = None,
         memory: Union[MemoryInterface, None] = None,
     ) -> None:
@@ -39,6 +41,7 @@ class HTTPTarget(PromptTarget):
         self.http_request = http_request
         self.callback_function = callback_function
         self.prompt_regex_string = prompt_regex_string
+        self.use_tls = use_tls
 
         if not self.http_request:
             raise ValueError("HTTP Request is required for HTTP Target")
@@ -63,7 +66,7 @@ class HTTPTarget(PromptTarget):
 
         # Add Prompt into request body (if the body takes it)
         if re.search(self.prompt_regex_string, http_body):
-            http_body = re_pattern.sub(request.converted_value, http_body)
+            http_body = re_pattern.sub(request.converted_value, http_body)  #
 
         response = requests.request(
             method=http_method,
@@ -130,10 +133,11 @@ class HTTPTarget(PromptTarget):
         http_url_beg = ""
         if len(http_req_info_line) > 2:
             http_version = http_req_info_line[2]
-            if "HTTP/2" in http_version:
-                http_url_beg = "https://"
-            elif "HTTP/1.1" in http_version:
-                http_url_beg = "http://"
+            if "HTTP/2" in http_version or "HTTP/1.1" in http_version:
+                if self.use_tls == True:
+                    http_url_beg = "https://"
+                else:
+                    http_url_beg = "http://"
             else:
                 raise ValueError(f"Unsupported protocol: {http_version}")
 
@@ -151,75 +155,3 @@ class HTTPTarget(PromptTarget):
 
         if len(request_pieces) != 1:
             raise ValueError("This target only supports a single prompt request piece.")
-
-
-def parse_json_factory(key: str) -> Callable:
-    """
-    Purpose: determines proper parsing response function for an HTTP Request
-    Parameters:
-        key (str): this is the path pattern to follow for parsing the output response
-            (ie for AOAI this would be choices[0].message.content)
-            (for BIC this needs to be a regex pattern for the desired output)
-        response_type (ResponseType): this is the type of response (ie HTML or JSON)
-    Returns: proper output parsing response
-    """
-
-    def parse_json_http_response(response: requests.Response):
-        """
-        Purpose: parses json outputs
-        Parameters:
-            response (response): the HTTP Response to parse
-        Returns: parsed output from response given a "key" path to follow
-        """
-        json_response = json.loads(response.content)
-        data_key = _fetch_key(data=json_response, key=key)
-        return data_key
-
-    return parse_json_http_response
-
-
-def parse_using_regex_text_factory(key: str, url: str = None) -> Callable:
-    def parse_using_regex(response: requests.Response):
-        """
-        Purpose: parses text outputs using regex
-        Parameters:
-            url (optional str): the original URL if this is needed to get a full URL response back (ie BIC)
-            key (str): this is the regex pattern to follow for parsing the output response
-            response (response): the HTTP Response to parse
-        Returns: parsed output from response given a regex pattern to follow
-        """
-        re_pattern = re.compile(key)
-        match = re.search(re_pattern, str(response.content))
-        if match:
-            if url:
-                return url + match.group()
-            else:
-                return match.group()
-        else:
-            return str(response.content)
-
-    return parse_using_regex
-
-
-def _fetch_key(data: dict, key: str) -> str:
-    """
-    Credit to @Mayuraggarwal1992
-    Fetches the answer from the HTTP JSON response based on the path.
-
-    Args:
-        data (dict): HTTP response data.
-        key (str): The key path to fetch the value.
-
-    Returns:
-        str: The fetched value.
-    """
-    pattern = re.compile(r"([a-zA-Z_]+)|\[(\d+)\]")
-    keys = pattern.findall(key)
-    for key_part, index_part in keys:
-        if key_part:
-            data = data.get(key_part, None)
-        elif index_part and isinstance(data, list):
-            data = data[int(index_part)] if len(data) > int(index_part) else None
-        if data is None:
-            return ""
-    return data
